Convert ext/xsl to HNI.
