+++ hphp/test/tools/import_zend_test.py
@@ -749,6 +749,9 @@ other_files = (
     '/ext/xmlreader/tests/relaxNG.rng',
     '/ext/xmlreader/tests/relaxNG2.rng',
     '/ext/xmlreader/tests/relaxNG3.rng',
+    '/ext/xsl/tests/53965/collection.xml',
+    '/ext/xsl/tests/53965/collection.xsl',
+    '/ext/xsl/tests/53965/include.xsl',
     '/ext/xsl/tests/area_list.xsl',
     '/ext/xsl/tests/area_name.xml',
     '/ext/xsl/tests/bug49634.xml',
