+++ hphp/test/zend/bad/ext/xsl/tests/bug53965.php.expectf
@@ -1,0 +1,4 @@
+Hey! Welcome to Nicolas Eliaszewicz's sweet CD collection! 
+  
+ <h1>Fight for your mind</h1><h2>by Ben Harper - 1995</h2><hr>
+ <h1>Electric Ladyland</h1><h2>by Jimi Hendrix - 1997</h2><hr>
