+++ hphp/test/zend/bad/ext/xsl/tests/bug53965.php.skipif
@@ -1,0 +1,3 @@
+<?php
+if (!extension_loaded('xsl')) die("skip Extension XSL is required\n");
+?>
