+++ hphp/test/zend/good/ext/xsl/tests/53965/collection.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<collection>
+ <cd>
+  <title>Fight for your mind</title>
+  <artist>Ben Harper</artist>
+  <year>1995</year>
+ </cd>
+ <cd>
+  <title>Electric Ladyland</title>
+  <artist>Jimi Hendrix</artist>
+  <year>1997</year>
+ </cd>
+</collection>
