+++ hphp/test/zend/good/ext/xsl/tests/53965/collection.xsl
@@ -1,0 +1,11 @@
+<xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
+
+ <xsl:include href="include.xsl"/>
+
+ <xsl:param name="owner" select="'Nicolas Eliaszewicz'"/>
+ <xsl:output method="html" encoding="iso-8859-1" indent="no"/>
+ <xsl:template match="collection">
+  Hey! Welcome to <xsl:value-of select="$owner"/>'s sweet CD collection! 
+  <xsl:apply-templates/>
+ </xsl:template>
+</xsl:stylesheet>
