+++ hphp/test/zend/good/ext/xsl/tests/xslt010.php.skipif
@@ -1,0 +1,6 @@
+<?php 
+require_once dirname(__FILE__) .'/skipif.inc';
+$proc = new xsltprocessor;
+if (!$proc->hasExsltSupport()) die('skip EXSLT support not available');
+if (LIBXSLT_VERSION < 10117) die('skip too old libxsl');
+?>
